LayoutTests/fast/scrolling/mac/scrollend-event-scroll-snap-expected.txt


Test scroll over content
PASS scrollendEventFired is true
PASS scrollendEventWasFiredDuringScroll is false
PASS successfullyParsed is true

TEST COMPLETE
